Add Failf for errors that must quit
Uses fprintf_s instead of copied code since it's simpler and we're sure
it works correctly.

Newlines have to have added manually, but at least it allows for better
control if ever we don't want a new line.

<a_string>.c_str() must be used for printf to show strings correctly,
maybe a better way of doing it so that we don't need to do this all the
time.